Support computation of stress.
We do this by replacing the displacment function with a
displacement under strain, i.e. (1 + epsilon) @ displacement,
and then wrapping the energy function so it accepts the strain
as argument. THEN we wrap it again to automatically pass the
strain, and do the derivative, and BAM.

I cannot *believe* this works. It doesn't even seem to be
particularly slow!